Several 541 fixes

+++ install_rstudio.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 set -ex
 
+# Add anaconda-project to PATH if not present
+source /opt/continuum/anaconda/bin/activate root
+if ! which anaconda-project; then
+  ln -s /opt/continuum/anaconda/envs/lab_launch/bin/anaconda-project /usr/local/bin/
+fi
+
 # Install psmisc for 5.4.x
 if [ ! $(rpm -qa psmisc) ]; then
    if [ ! -f psmisc-22.20-16.el7.x86_64.rpm ]; then
@@ -10,20 +16,16 @@ if [ ! $(rpm -qa psmisc) ]; then
 fi
 
 # Install RStudio server
-if [ ! -f rstudio-server-rhel-1.2.5033-x86_64.rpm ]; then
-   curl -O https://download2.rstudio.org/server/centos6/x86_64/rstudio-server-rhel-1.2.5033-x86_64.rpm
+if [ ! -f rstudio-server-rhel-1.2.5042-x86_64.rpm ]; then
+   curl -O https://download2.rstudio.org/server/centos6/x86_64/rstudio-server-rhel-1.2.5042-x86_64.rpm
 fi
-rpm -i rstudio-server-rhel-1.2.5033-x86_64.rpm
+rpm -i rstudio-server-rhel-1.2.5042-x86_64.rpm
 rm *.rpm
 
-# for some reason on AE541 ae-editor, conda is not in path and build fail - using full path... 
-conda=conda
-[[ command -v $conda ]] || conda='/opt/continuum/anaconda/condabin/conda'
-[[ -f $conda ]] || exit 1
-
-# Create the custom environments - only if environment files are present 
-envs=$(ls *.txt 2>/dev/null)
-for envtxt in $envs; do
+# Create the custom environments
+for envtxt in *.txt; do
+    # if there are no text files, envtxt will actually be *.txt
+    [ -e "$envtxt" ] || continue
     envname=${envtxt%.txt}
     CONDARC=./condarc conda create -n $envname --file $envtxt
     [ -d /opt/continuum/anaconda/envs/$envname/conda-meta ] || exit -1
@@ -37,17 +39,15 @@ if [ $envname ]; then
     conda clean --all
 fi
 
-
 # 5.3.x back compatibility fixes
 if [ ! -f /opt/continuum/scripts/start_user.sh ]; then
     cp startup.sh build_condarc.py run_tool.py /opt/continuum/scripts/
 fi
 
 # Rstudio scripts
-cp Rprofile /opt/continuum/.Rprofile
-cp rsession.sh start_rstudio.sh /opt/continuum/scripts/
-
-# Fix ownership and permissions
-chmod +x /opt/continuum/scripts/*.sh
-chown anaconda:anaconda /opt/continuum/.Rprofile /opt/continuum/scripts/*.sh
-
+if [ ! -f /opt/continuum/.Rprofile ]; then
+    cp Rprofile /opt/continuum/.Rprofile
+    cp rsession.sh start_rstudio.sh /opt/continuum/scripts/
+    chmod +x /opt/continuum/scripts/*.sh
+    chown anaconda:anaconda /opt/continuum/.Rprofile /opt/continuum/scripts/*.sh
+fi
